refactor!: give builder, provers and verifiers more descriptive names
