+++ co-noir/co-builder/src/prelude.rs
@@ -1,5 +1,4 @@
 pub use crate::acir_format::{AcirFormat, HonkRecursion};
-pub use crate::builder::{GenericUltraCircuitBuilder, UltraCircuitBuilder};
 pub use crate::crs::Crs;
 pub use crate::crs::ProverCrs;
 pub use crate::crs::parse::CrsParser;
@@ -16,5 +15,6 @@ pub use crate::types::types::{
     AGGREGATION_OBJECT_SIZE, ActiveRegionData, CycleNode, CyclicPermutation, NUM_SELECTORS,
     NUM_WIRES, PAIRING_POINT_ACCUMULATOR_SIZE, ZeroKnowledge,
 };
+pub use crate::ultra_builder::{GenericUltraCircuitBuilder, UltraCircuitBuilder};
 pub use crate::utils::Utils;
 pub use co_acvm::PlainAcvmSolver;
