+++ co-noir/co-builder/src/types/blake3.rs
@@ -3,9 +3,9 @@ use {
     super::blake2s::BLAKE2S_IV,
     super::plookup::{ColumnIdx, MultiTableId, Plookup},
     crate::TranscriptFieldType,
-    crate::builder::GenericUltraCircuitBuilder,
     crate::prelude::HonkCurve,
     crate::types::field_ct::{ByteArray, FieldCT},
+    crate::ultra_builder::GenericUltraCircuitBuilder,
     ark_ff::PrimeField,
     co_acvm::mpc::NoirWitnessExtensionProtocol,
     std::array,
