+++ co-noir/co-builder/src/types/blake_util.rs
@@ -1,4 +1,4 @@
-use crate::{TranscriptFieldType, builder::GenericUltraCircuitBuilder, prelude::HonkCurve};
+use crate::{TranscriptFieldType, prelude::HonkCurve, ultra_builder::GenericUltraCircuitBuilder};
 use ark_ec::pairing::Pairing;
 use ark_ff::PrimeField;
 use co_acvm::mpc::NoirWitnessExtensionProtocol;
