+++ co-noir/co-ultrahonk/src/co_decider/co_shplemini/mod.rs
@@ -1,2 +1,2 @@
-pub(crate) mod prover;
+pub(crate) mod co_shplemini_prover;
 pub(crate) mod types;
