+++ co-noir/co-ultrahonk/src/co_decider/mod.rs
@@ -1,6 +1,6 @@
+pub(crate) mod co_decider_prover;
 pub(crate) mod co_shplemini;
 pub(crate) mod co_sumcheck;
-pub(crate) mod prover;
 pub(crate) mod relations;
 pub(crate) mod small_subgroup_ipa;
 pub(crate) mod types;
