+++ co-noir/co-ultrahonk/src/prelude.rs
@@ -1,6 +1,6 @@
 pub use crate::key::proving_key::{ProvingKey, Rep3ProvingKey, ShamirProvingKey};
 // Re-exporting the following types from `ultrahonk` and `co_builder` crates:
-pub use crate::prover::{CoUltraHonk, Rep3CoUltraHonk, ShamirCoUltraHonk};
+pub use crate::co_ultra_prover::{CoUltraHonk, Rep3CoUltraHonk, ShamirCoUltraHonk};
 pub use crate::types::Polynomials;
 pub use crate::{
     PlainCoBuilder, Rep3CoBuilder, ShamirCoBuilder, mpc_prover_flavour::MPCProverFlavour,
@@ -12,7 +12,6 @@ pub use co_builder::prelude::{
 pub use co_builder::prelude::{Crs, CrsParser, HonkCurve, Polynomial, ProverCrs};
 pub use ultrahonk::Utils;
 pub use ultrahonk::prelude::UltraCircuitBuilder;
-pub use ultrahonk::prelude::UltraFlavour;
 pub use ultrahonk::prelude::UltraHonk;
 pub use ultrahonk::prelude::VerifyingKeyBarretenberg;
 pub use ultrahonk::prelude::ZeroKnowledge;
