+++ co-noir/ultrahonk/src/decider/decider_verifier.rs
@@ -1,7 +1,7 @@
 use super::types::VerifierMemory;
 use crate::CONST_PROOF_SIZE_LOG_N;
 use crate::NUM_LIBRA_COMMITMENTS;
-use crate::{Utils, plain_prover_flavour::PlainProverFlavour, verifier::HonkVerifyResult};
+use crate::{Utils, plain_prover_flavour::PlainProverFlavour, ultra_verifier::HonkVerifyResult};
 use ark_ec::AffineRepr;
 use ark_ff::{One, Zero};
 use co_builder::{
