+++ co-noir/ultrahonk/src/decider/mod.rs
@@ -1,9 +1,9 @@
 pub(crate) mod barycentric;
-pub(crate) mod prover;
+pub(crate) mod decider_prover;
+pub(crate) mod decider_verifier;
 pub(crate) mod relations;
 pub mod shplemini;
 pub(crate) mod small_subgroup_ipa;
 pub(crate) mod sumcheck;
 pub(crate) mod types;
 pub(crate) mod univariate;
-pub(crate) mod verifier;
