+++ co-noir/ultrahonk/src/decider/shplemini/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod prover;
+pub(crate) mod shplemini_prover;
+pub(crate) mod shplemini_verifier;
 pub(crate) mod types;
-pub(crate) mod verifier;
