+++ co-noir/ultrahonk/src/decider/shplemini/shplemini_verifier.rs
@@ -4,10 +4,10 @@ use crate::{
     CONST_PROOF_SIZE_LOG_N, NUM_INTERLEAVING_CLAIMS, NUM_LIBRA_COMMITMENTS,
     NUM_SMALL_IPA_EVALUATIONS,
     decider::{
+        decider_verifier::DeciderVerifier,
         types::{ClaimedEvaluations, VerifierCommitments},
-        verifier::DeciderVerifier,
     },
-    verifier::HonkVerifyResult,
+    ultra_verifier::HonkVerifyResult,
 };
 use ark_ec::AffineRepr;
 use ark_ff::{Field, One, Zero};
