+++ co-noir/ultrahonk/src/decider/sumcheck/sumcheck_verifier.rs
@@ -1,10 +1,12 @@
 use super::SumcheckVerifierOutput;
 use crate::CONST_PROOF_SIZE_LOG_N;
 use crate::{
-    decider::{sumcheck::round_verifier::SumcheckVerifierRound, verifier::DeciderVerifier},
+    decider::{
+        decider_verifier::DeciderVerifier, sumcheck::sumcheck_round_verifier::SumcheckVerifierRound,
+    },
     plain_prover_flavour::PlainProverFlavour,
     prelude::GateSeparatorPolynomial,
-    verifier::HonkVerifyResult,
+    ultra_verifier::HonkVerifyResult,
 };
 use ark_ff::{One, Zero};
 use co_builder::prelude::{HonkCurve, ZeroKnowledge};
