+++ co-noir/ultrahonk/src/oink/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod prover;
+pub(crate) mod oink_prover;
+pub(crate) mod oink_verifier;
 pub(crate) mod types;
-pub(crate) mod verifier;
