+++ co-noir/ultrahonk/src/prelude.rs
@@ -1,8 +1,7 @@
 pub use crate::decider::barycentric::Barycentric;
 pub use crate::decider::types::GateSeparatorPolynomial;
 pub use crate::decider::univariate::Univariate;
-pub use crate::prover::UltraHonk;
-pub use co_builder::flavours::ultra_flavour::UltraFlavour;
+pub use crate::ultra_prover::UltraHonk;
 pub use co_builder::prelude::PlainAcvmSolver;
 pub use co_builder::prelude::VerifyingKeyBarretenberg;
 pub use co_builder::prelude::{ProvingKey, UltraCircuitBuilder, ZeroKnowledge};
