+++ co-noir/ultrahonk/src/ultra_verifier.rs
@@ -1,8 +1,8 @@
 use crate::{
-    decider::{types::VerifierMemory, verifier::DeciderVerifier},
-    oink::verifier::OinkVerifier,
+    decider::{decider_verifier::DeciderVerifier, types::VerifierMemory},
+    oink::oink_verifier::OinkVerifier,
     plain_prover_flavour::PlainProverFlavour,
-    prover::UltraHonk,
+    ultra_prover::UltraHonk,
 };
 use co_builder::prelude::{HonkCurve, VerifyingKey, ZeroKnowledge};
 use common::HonkProof;
